requirements.txt
flask==2.0.1
python-telegram-bot==20.1
requests==2.28.2